# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API'
  description: 'API for getting and managing of Rick&Morty characters'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/registro:
    post:
      summary: Registro
      operationId: registro
      description: 'Este metodo registra al usuario en el API'
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "User's name."
                  example: user1
                email:
                  type: string
                  description: "User's email."
                  example: test@test.com
                password:
                  type: string
                  description: "User's password."
                  example: '12345'
                password_confirmation:
                  type: string
                  description: "User's password for confirmation."
                  example: '12345'
              required:
                - name
                - email
                - password
                - password_confirmation
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []